+++ packages/vite/src/node/ssr/__tests__/ssrTransform.spec.ts
@@ -119,6 +119,15 @@ test('export * from', async () => {
   `)
 })
 
+test('export * as from', async () => {
+  expect((await ssrTransform(`export * as foo from 'vue'`, null, null)).code)
+    .toMatchInlineSnapshot(`
+    "const __vite_ssr_import_0__ = __vite_ssr_import__(\\"vue\\")
+
+    Object.defineProperty(__vite_ssr_exports__, \\"foo\\", { enumerable: true, configurable: true, get(){ return __vite_ssr_import_0__ }})"
+  `)
+})
+
 test('export default', async () => {
   expect(
     (await ssrTransform(`export default {}`, null, null)).code
@@ -224,7 +233,8 @@ test('overwrite bindings', async () => {
           `function c() { const { test: inject } = { test: true }; console.log(inject) }\n` +
           `const d = inject \n` +
           `function f() {  console.log(inject) }\n` +
-          `function e() { const { inject } = { inject: true } }\n`,
+          `function e() { const { inject } = { inject: true } }\n` +
+          `function g() { const f = () => { const inject = true }; console.log(inject) }\n`,
         null,
         null
       )
@@ -237,6 +247,7 @@ test('overwrite bindings', async () => {
     const d = __vite_ssr_import_0__.inject 
     function f() {  console.log(__vite_ssr_import_0__.inject) }
     function e() { const { inject } = { inject: true } }
+    function g() { const f = () => { const inject = true }; console.log(__vite_ssr_import_0__.inject) }
     "
   `)
 })